Module Curve_sampling.P2

Interface using Gg.p2 to represent points.

val param : ?n:int -> ?viewport:Gg.Box2.t -> ?init:float list -> ?init_pt:(float * Gg.p2) list -> (float -> Gg.p2) -> float -> float -> [ `Fn ] t

See Curve_sampling.param.

val uniform : ?n:int -> (float -> Gg.p2) -> float -> float -> [ `Fn ] t

uniform f a b return a sampling of the image of f on n equidistant points in the interval [a, b] (the boundaries a and b being always included — so n >= 2).

parameter n
the number of points. If n <= 2 is given, it is considered as if n=2 was passed. Default: n = 100.

val of_path : Gg.p2 list -> [ `Pt ] t

Use the provided path as the sampling.

type point_or_cut =

| Point of Gg.p2

| Cut

val to_list : _ t -> point_or_cut list

to_list s return the sampling as a list of points in increasing order of the parameter of the curve. The curve is possibly made of several pieces separated by a single Cut.

val of_seq : ?n:int -> Gg.p2 Stdlib.Seq.t -> [ `Pt ] t

See Curve_sampling.of_seq.

val to_seq : _ t -> point_or_cut Stdlib.Seq.t

See Curve_sampling.to_seq.